Renamed the example program.
