+++ examples/WINDOWS_multi_feature/WINDOWS_multi_feature.cpp
@@ -1,4 +1,4 @@
-#include "WIN32_static_helper.hpp"
+#include "WINDOWS_multi_feature.hpp"
 #include "antirtos.h"
 #include <iostream>
