+++ examples/WINDOWS_multi_feature/WINDOWS_multi_feature.hpp
